[MONDRIAN-1769] First pass at creating the extension points for non-sql support. The places that need cleanup are all marked with 'FIXME MONGO'.

+++ src/main/mondrian/rolap/agg/Aggregation.java
@@ -140,6 +140,7 @@ public class Aggregation {
         if (groupingSetsCollector.useGroupingSets()) {
             groupingSetsCollector.add(groupingSet);
         } else {
+            // FIXME MONGO Make this pluggable.
             final SegmentLoader segmentLoader = new SegmentLoader(cacheMgr);
             segmentLoader.load(
                 cellRequestCount,
@@ -248,7 +249,7 @@ public class Aggregation {
             }
 
             // more than one - check for children of same parent
-            double constraintLength = (double) valueCount;
+            double constraintLength = valueCount;
             Member parent = null;
             Level level = null;
             for (int j = 0; j < valueCount; j++) {
